# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Believer Backend API
  version: 1.8.0
  description: Active REST API contract for auth, account recovery, account-governance settings actions, business listings, mosques, backend-owned mosque prayer times, mosque page content, mosque image uploads, bookmarks, notifications, and services.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request payload
        details:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          type: object
          example: {}
...
    SuccessMessageEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          required: [success, message]
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: If an account exists for that email, a password reset link has been sent.
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
...
    DeactivateAccountRequest:
      type: object
      required: [confirmation]
      properties:
        confirmation:
          type: string
          enum: [DEACTIVATE]
    SupportRequest:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
          minLength: 4
          maxLength: 120
        message:
          type: string
          minLength: 10
          maxLength: 4000
    MosqueSuggestionRequest:
      type: object
      required: [mosqueName, city, country]
      properties:
        mosqueName:
          type: string
          minLength: 2
          maxLength: 180
        city:
          type: string
          minLength: 2
          maxLength: 120
        country:
          type: string
          minLength: 2
          maxLength: 120
        addressLine:
          type: string
          maxLength: 240
        notes:
          type: string
          maxLength: 2000
...
paths:
...
  /api/v1/auth/deactivate:
    post:
      tags: [Auth]
      summary: Deactivate the authenticated account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeactivateAccountRequest'
      responses:
        '200':
          description: Account deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/account/support-requests:
    post:
      tags: [Account Governance]
      summary: Submit a support/contact request for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportRequest'
      responses:
        '201':
          description: Support request stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/account/mosque-suggestions:
    post:
      tags: [Account Governance]
      summary: Submit a mosque suggestion for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MosqueSuggestionRequest'
      responses:
        '201':
          description: Mosque suggestion stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...